+++ src/qcommon/cvar.c
@@ -2068,6 +2068,88 @@ Cvar_Restart(qbool unsetVM)
   }
 }
 
+static void
+Cvar_Trim(qbool verbose)
+{
+  cvar_t *curvar = cvar_vars;
+
+  while(curvar)
+  {
+    if (curvar->flags & CVAR_USER_CREATED)
+    {
+      //throw out any variables the user created
+      if (verbose)
+      {
+        Com_Printf("unset cvar" S_COLOR_YELLOW " %s\n", curvar->name);
+      }
+
+      curvar = Cvar_Unset(curvar);
+      continue;
+    }
+
+    curvar = curvar->next;
+  }
+}
+
+/*
+============
+Cvar_Trim_f
+
+Removes all user-created cvars
+This will only accept to run when both the server and client are running unless forced
+============
+*/
+static void
+Cvar_Trim_f(void)
+{
+  qbool forced = qfalse;
+  qbool verbose = qtrue;
+  qint i;
+
+  for(i = 1;i < Cmd_Argc();i++)
+  {
+    const qchar *s = Cmd_Argv(i);
+
+    if (*s == '-')
+    {
+      s++;
+
+      while(*s != '\0')
+      {
+        if (*s == 'f') //force cleanup
+        {
+          forced = qtrue;
+        }
+        else if (*s == 's') //silent mode
+        {
+          verbose = qfalse;
+        }
+
+        s++;
+      }
+    }
+  }
+
+#if defined(DEDICATED)
+  if ((com_sv_running && com_sv_running->integer) || forced)
+#else
+  if ((com_cl_running && com_cl_running->integer && com_sv_running && com_sv_running->integer) || forced)
+#endif
+  {
+    Cvar_Trim(verbose);
+    return;
+  }
+
+#if defined(DEDICATED)
+  Com_Printf(S_COLOR_YELLOW " You're not running a server, so not all subsystems/VMs are loaded.\n");
+#else
+  Com_Printf(S_COLOR_YELLOW " You're not running a listen server, so not all subsystems/VMs are loaded.\n");
+#endif
+  Com_Printf(S_COLOR_YELLOW " This means you'd remove cvars that are probably best kept around.\n");
+  Com_Printf(S_COLOR_YELLOW " If you don't care, you can force the call by running '\\%s -f'.\n", Cmd_Argv(0));
+  Com_Printf(S_COLOR_YELLOW " You've been warned.\n");
+}
+
 /*
 ============
 Cvar_Restart_f
@@ -2455,4 +2537,5 @@ Cvar_Init(void)
   Cmd_AddCommand("cvarlist", Cvar_List_f);
   Cmd_AddCommand("cvar_modified", Cvar_ListModified_f);
   Cmd_AddCommand("cvar_restart", Cvar_Restart_f);
+  Cmd_AddCommand("cvar_trim", Cvar_Trim_f);
 }